refactor(cli): extract runtime plan handler
